pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "k8s_manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "kubernetes>=32.0.1",
    "mcp[cli]>=1.3.0",
    "numpy>=2.2.3",
    "pick>=2.4.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "statsmodels>=0.14.4",
]

[tool.uv.workspace]
members = ["k8s_manager"]

[tool.setuptools]
package-dir = {"" = "src"}

[project.scripts]
k8s_manager = "k8s_manager:main"
